feat(quality): TASK-47 — Ruff plumbing (Option 1 rule-set tuning) [skip-simplify]
Lands Ruff as the Python lint + format + import-sort tool. Hard-fail in
CI via tests.yml (job: ruff). Config in pyproject.toml [tool.ruff].

Unblocked by human picking Option 1 from the BLOCKED note:
- extend-immutable-calls = ["fastapi.Depends", "fastapi.Query",
  "fastapi.Path", "fastapi.Body", "fastapi.Header", "fastapi.Form",
  "fastapi.File", "fastapi.Cookie"] — kills the 92 B008 false-positives
  on FastAPI's dependency-injection pattern.
- ignore = ["DTZ011", "RUF001", "RUF002", "RUF003"] — date.today() is
  valid for naive-date contexts; ambiguous-unicode rules false-positive
  on NRL content (×, —, ° in match reports / operator notes).
- line-length = 120 (up from 100; matches the codebase's existing
  style).
- per-file-ignores E501 for 8 prompt / HTML-content files (LLM system
  prompts, SES email bodies, seed-mock-articles markdown) where line
  breaks damage the artifact.

Source-tree diff: ~190 files (4730/-3666). Composition:
- 337 safe autofixes via `ruff check --fix` (import sort, datetime
  modernization, unused imports, etc.).
- 24 unsafe autofixes via `--unsafe-fixes` (RUF046 int(round(...)) →
  round(...); RUF059 _-prefix on unused-unpacked vars).
- 142 files reformatted via `ruff format` (line wrapping, trailing
  commas, spacing normalization).
- Manual fixes:
  * F821 typo in worker-publishing/.../update_consensus.py:125
    (entity_claims → person_claims) — REAL BUG Ruff caught; the
    NameError would have fired if the snapshot path was ever entered.
  * 33 B904 raise-from-inside-except — batch-fixed via a one-shot
    helper script that found the enclosing `except ... as Y:` and
    appended `from Y`. The 2 unbound `except ValueError:` cases
    (insights.py:84, lineup.py:288) hand-fixed to `from None`.
  * 2 B008 `# noqa` markers on `VoiceClusterParams()` defaults
    (verified frozen dataclass — immutable).
  * 1 B007 in reconcile_orphan_channels.py (url → _url).
  * 1 RUF034 no-op ternary in scraping/nrl.py.
  * 5 E402 `# noqa` markers in scripts with sys.path.insert before
    imports.
  * 10 E501 inline `# noqa` markers on SQL CHECK / log format / f-string
    error message lines where wrapping hurts readability.

Local proof: make lint-python exits 0 (All checks passed!); make
typecheck-python exits 0; pytest tests/unit → 398 passed (no
regression).

META.md: new `### Datetime / timezone` invariant subsection. Uses
`datetime.now(UTC)` form (matches reality post UP017 autofix).

adversarial-reviewer: PASS WITH CONCERNS. Per-file ignore scopes,
inline noqa rationales, F821 fix, B904 batch-fix, and B008 frozen-
dataclass justification all spot-checked correct.

Per PLAN.md "Engineering quality hardening — Tier 1".

+++ packages/shared/jeromelu_shared/players/nrlcom.py
@@ -27,7 +27,6 @@ from typing import Any
 
 import httpx
 
-
 PROFILE_URL = "https://www.nrl.com/players/nrl-premiership/{team_short}/{slug}/"
 USER_AGENT = "Mozilla/5.0 (compatible; jeromelu-roster/1.0)"
 
@@ -40,13 +39,14 @@ _JSONLD_RE = re.compile(
 @dataclass
 class NrlProfile:
     """Subset of the JSON-LD ``Person`` block we promote into the DB."""
+
     canonical_name: str
     dob: date | None
-    birthplace_text: str | None     # raw e.g. "Sydney, NSW" — no normalisation in v1
+    birthplace_text: str | None  # raw e.g. "Sydney, NSW" — no normalisation in v1
     image_url: str | None
     height_cm: int | None
     weight_kg: int | None
-    job_title: str | None           # raw e.g. "Captain - Halfback" — captaincy ignored
+    job_title: str | None  # raw e.g. "Captain - Halfback" — captaincy ignored
 
 
 _APOSTROPHE_OR_DOT_RE = re.compile(r"[\.‘’']")
@@ -78,7 +78,7 @@ def slugify_name(name: str) -> str:
 # against the live league — see comments in
 # `docs/agents/system/player-roster.md`.
 NRLCOM_TEAM_OVERRIDES: dict[str, str] = {
-    "tigers":    "wests-tigers",
+    "tigers": "wests-tigers",
     "rabbitohs": "south-sydney-rabbitohs",
 }
 
@@ -183,6 +183,6 @@ def _qv(node: Any, *, expected_unit: str) -> int | None:
     if v is None:
         return None
     try:
-        return int(round(float(v)))
+        return round(float(v))
     except (TypeError, ValueError):
         return None
